backend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
Pillow
Python-multipart
scikit-learn